requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
APScheduler>=3.10.4
instagrapi>=2.3.0
